+++ Tests/Collection/Immutable/Generic/ListTest.php
@@ -320,6 +320,16 @@ class ListTest extends TestCase
         $this->assertEquals(10 + 1 + 2 + 3, $list->reduce('($t, $v) => $t + $v', 10));
     }
 
+    public function testShouldReduceListWithInitialValueToOtherType()
+    {
+        $list = new ListCollection('int');
+        $list = $list->add(1);
+        $list = $list->add(2);
+        $list = $list->add(3);
+
+        $this->assertEquals('123', $list->reduce('($t, $v) => $t . $v', ''));
+    }
+
     public function testShouldClearCollection()
     {
         $this->list = $this->list->add('value');